chore: remove builder circular dep
