+++ packages/runtime/src/builder/serialization/control-serialization.ts
@@ -1,5 +1,4 @@
 import {
-  ComboboxControlDefinition,
   ComboboxControlType,
   ListControlDefinition,
   ListControlType,
@@ -12,90 +11,121 @@ import {
 } from '../../controls'
 import {
   CheckboxDescriptor as CheckboxControl,
-  CheckboxValue as CheckboxControlValue,
   Data,
-  Device,
   DateDescriptor as DateControl,
-  DateValue as DateControlValue,
-  Gap,
   GapXDescriptor as GapXControl,
-  GapXValue as GapXControlValue,
   GapYDescriptor as GapYControl,
-  GapYValue as GapYControlValue,
   ImageDescriptor as ImageControl,
-  ImageValue as ImageControlValue,
   LinkDescriptor as LinkControl,
-  LinkValue as LinkControlValue,
   ListDescriptor as ListControl,
-  ListOptions as ListControlConfig,
   ListValue as ListControlValue,
   NumberDescriptor as NumberControl,
-  NumberValue as NumberControlValue,
   ResponsiveColorDescriptor as ResponsiveColorControl,
-  ResponsiveColorValue as ResponsiveColorControlValue,
   ResponsiveNumberDescriptor as ResponsiveNumberControl,
-  ResponsiveNumberValue as ResponsiveNumberControlValue,
   ResponsiveIconRadioGroupDescriptor as ResponsiveIconRadioGroupControl,
-  ResponsiveIconRadioGroupValue as ResponsiveIconRadioGroupControlValue,
   ResponsiveSelectDescriptor as ResponsiveSelectControl,
-  ResponsiveSelectValue as ResponsiveSelectControlValue,
   ResponsiveLengthDescriptor as ResponsiveLengthControl,
-  ResponsiveLengthValue as ResponsiveLengthControlValue,
   RichTextDescriptor as RichTextControl,
-  RichTextValue as RichTextControlValue,
   ShapeDescriptor as ShapeControl,
-  ShapeValue as ShapeControlValue,
   TextInputDescriptor as TextInputControl,
-  TextInputValue as TextInputControlValue,
   TextStyleDescriptor as TextStyleControl,
-  TextStyleValue as TextStyleControlValue,
   TypeaheadDescriptor as TypeaheadControl,
-  TypeaheadOptions as TypeaheadControlConfig,
   TypeaheadValue as TypeaheadControlValue,
   PanelDescriptor as PanelControl,
-  PanelDescriptorType as PanelControlType,
   PanelDescriptorValueType as PanelControlValueType,
   PropControllerDescriptor as Control,
   Props as Controls,
 } from '../../prop-controllers'
 import {
-  IconRadioGroupOption,
-  LengthOption,
-  SelectLabelOrientation,
-  SelectOption,
-  Length,
-} from '../../prop-controllers/descriptors'
-import {
   deserializeComboboxControlDefinition,
   serializeComboboxControlDefinition,
 } from './controls/combobox'
-import { deserializeListControlDefinition, serializeListControlDefinition } from './controls/list'
-import { deserializeRichTextControlV2, serializeRichTextControlV2 } from './controls/rich-text-v2'
-import {
-  deserializeShapeControlDefinition,
-  serializeShapeControlDefinition,
-} from './controls/shape'
-import { deserializeStyleV2Control, serializeStyleV2Control } from './controls/style-v2'
 import { Deserialize, Serialize } from './controls/types'
 import {
-  DeserializedFunction,
   deserializeFunction,
   isSerializedFunction,
-  SerializedFunction,
   serializeFunction,
 } from './function-serialization'
+import {
+  DeserializedCheckboxControl,
+  DeserializedControl,
+  DeserializedDateControl,
+  DeserializedGapXControl,
+  DeserializedGapYControl,
+  DeserializedImageControl,
+  DeserializedLinkControl,
+  DeserializedListControl,
+  DeserializedNumberControl,
+  DeserializedPanelControl,
+  DeserializedResponsiveColorControl,
+  DeserializedResponsiveIconRadioGroupControl,
+  DeserializedResponsiveLengthControl,
+  DeserializedResponsiveNumberControl,
+  DeserializedResponsiveSelectControl,
+  DeserializedRichTextControl,
+  DeserializedShapeControl,
+  DeserializedTextInputControl,
+  DeserializedTextStyleControl,
+  DeserializedTypeaheadControl,
+  SerializedCheckboxControl,
+  SerializedControl,
+  SerializedDateControl,
+  SerializedGapXControl,
+  SerializedGapYControl,
+  SerializedImageControl,
+  SerializedLinkControl,
+  SerializedListControl,
+  SerializedNumberControl,
+  SerializedPanelControl,
+  SerializedPanelControlValueType,
+  SerializedResponsiveColorControl,
+  SerializedResponsiveIconRadioGroupControl,
+  SerializedResponsiveLengthControl,
+  SerializedResponsiveNumberControl,
+  SerializedResponsiveSelectControl,
+  SerializedRichTextControl,
+  SerializedShapeControl,
+  SerializedTextInputControl,
+  SerializedTextStyleControl,
+  SerializedTypeaheadControl,
+} from './types'
 
-type SerializedShapeControlConfig<T extends Record<string, SerializedPanelControl>> = {
-  type: T
-  preset?: { [K in keyof T]?: SerializedPanelControlValueType<T[K]> }
+function serializeShapeControlDefinition(
+  definition: ShapeControlDefinition,
+): [Serialize<ShapeControlDefinition>, Transferable[]] {
+  const [type, transferables] = Object.entries(definition.config.type).reduce(
+    ([type, transferables], [key, value]) => {
+      const [serializedType, serializedTransferables] = serializeControl(value)
+
+      type[key] = serializedType
+      transferables.push(...serializedTransferables)
+
+      return [type, transferables]
+    },
+    [{} as Record<string, SerializedControl>, [] as Transferable[]],
+  )
+
+  return [{ ...definition, config: { ...definition.config, type } }, transferables] as [
+    Serialize<ShapeControlDefinition>,
+    Transferable[],
+  ]
 }
 
-type SerializedShapeControl<
-  _T extends Record<string, Data>,
-  U extends Record<string, SerializedPanelControl>,
-> = {
-  type: typeof Controls.Types.Shape
-  options: SerializedShapeControlConfig<U>
+function deserializeShapeControlDefinition(
+  definition: Serialize<ShapeControlDefinition>,
+): Deserialize<Serialize<ShapeControlDefinition>> {
+  const type = Object.entries(definition.config.type).reduce((type, [key, value]) => {
+    const serializedType = deserializeControl(value)
+
+    type[key] = serializedType
+
+    return type
+  }, {} as Record<string, DeserializedControl>)
+
+  return {
+    ...definition,
+    config: { ...definition.config, type },
+  } as Deserialize<Serialize<ShapeControlDefinition>>
 }
 
 function serializeShapeControl<
@@ -127,19 +157,6 @@ function serializeShapeControl<
   return [{ ...control, options: { ...control.options, type: serializedType } }, transferables]
 }
 
-type DeserializedShapeControlConfig<T extends Record<string, DeserializedPanelControl>> = {
-  type: T
-  preset?: { [K in keyof T]?: DeserializedPanelControlValueType<T[K]> }
-}
-
-type DeserializedShapeControl<
-  _T extends Record<string, Data>,
-  U extends Record<string, DeserializedPanelControl>,
-> = {
-  type: typeof Controls.Types.Shape
-  options: DeserializedShapeControlConfig<U>
-}
-
 function deserializeShapeControl<
   T extends Record<string, Data>,
   U extends Record<string, SerializedPanelControl>,
@@ -164,17 +181,32 @@ function deserializeShapeControl<
   return { ...control, options: { ...control.options, type: deserializedType } }
 }
 
-type SerializedListControlConfig<T extends Data> = {
-  type: SerializedPanelControl<T>
-  label?: string
-  getItemLabel?: SerializedFunction<Exclude<ListControlConfig<T>['getItemLabel'], undefined>>
-  preset?: ListControlValue<T>
-  defaultValue?: ListControlValue<T>
+function serializeListControlDefinition(
+  definition: ListControlDefinition,
+): [Serialize<ListControlDefinition>, Transferable[]] {
+  const [type, transferables] = serializeControl(definition.config.type)
+  const getItemLabel =
+    definition.config.getItemLabel && serializeFunction(definition.config.getItemLabel)
+
+  if (getItemLabel) transferables.push(getItemLabel)
+
+  return [
+    { ...definition, config: { ...definition.config, type, getItemLabel } },
+    transferables,
+  ] as [Serialize<ListControlDefinition>, Transferable[]]
 }
 
-type SerializedListControl<T extends ListControlValue = ListControlValue> = {
-  type: typeof Controls.Types.List
-  options: SerializedListControlConfig<T extends ListControlValue<infer U> ? U : never>
+function deserializeListControlDefinition(
+  definition: Serialize<ListControlDefinition>,
+): Deserialize<Serialize<ListControlDefinition>> {
+  const type = deserializeControl(definition.config.type)
+  const getItemLabel =
+    definition.config.getItemLabel && deserializeFunction(definition.config.getItemLabel)
+
+  return {
+    ...definition,
+    config: { ...definition.config, type, getItemLabel },
+  } as Deserialize<Serialize<ListControlDefinition>>
 }
 
 function serializeListControl<T extends Data>(
@@ -202,19 +234,6 @@ function serializeListControl<T extends Data>(
   ]
 }
 
-type DeserializedListControlConfig<T extends Data> = {
-  type: DeserializedPanelControl<T>
-  label?: string
-  getItemLabel?: DeserializedFunction<Exclude<ListControlConfig<T>['getItemLabel'], undefined>>
-  preset?: ListControlValue<T>
-  defaultValue?: ListControlValue<T>
-}
-
-type DeserializedListControl<T extends ListControlValue = ListControlValue> = {
-  type: typeof Controls.Types.List
-  options: DeserializedListControlConfig<T extends ListControlValue<infer U> ? U : never>
-}
-
 function deserializeListControl<T extends Data>(
   serializedControl: SerializedListControl<ListControlValue<T>>,
 ): DeserializedListControl<ListControlValue<T>> {
@@ -233,18 +252,6 @@ function deserializeListControl<T extends Data>(
   }
 }
 
-type SerializedTypeaheadControlConfig<T extends Data> = {
-  getItems: SerializedFunction<TypeaheadControlConfig<T>['getItems']>
-  label?: string
-  preset?: TypeaheadControlValue<T>
-  defaultValue?: TypeaheadControlValue<T>
-}
-
-type SerializedTypeaheadControl<T extends TypeaheadControlValue = TypeaheadControlValue> = {
-  type: typeof Controls.Types.Typeahead
-  options: SerializedTypeaheadControlConfig<T['value']>
-}
-
 function serializeTypeaheadControl<T extends Data>(
   control: TypeaheadControl<TypeaheadControlValue<T>>,
 ): [SerializedTypeaheadControl<TypeaheadControlValue<T>>, Transferable[]] {
@@ -256,18 +263,6 @@ function serializeTypeaheadControl<T extends Data>(
     { ...control, options: { ...control.options, getItems: serializedGetItems } },
     serializedGetItems == null ? [] : [serializedGetItems],
   ]
-}
-
-type DeserializedTypeaheadControlConfig<T extends Data> = {
-  getItems: DeserializedFunction<TypeaheadControlConfig<T>['getItems']>
-  label?: string
-  preset?: TypeaheadControlValue<T>
-  defaultValue?: TypeaheadControlValue<T>
-}
-
-type DeserializedTypeaheadControl<T extends TypeaheadControlValue = TypeaheadControlValue> = {
-  type: typeof Controls.Types.Typeahead
-  options: DeserializedTypeaheadControlConfig<T['value']>
 }
 
 function deserializeTypeaheadControl<T extends Data>(
@@ -283,29 +278,6 @@ function deserializeTypeaheadControl<T extends Data>(
   }
 }
 
-type SerializedConfig<T> =
-  | T
-  | SerializedFunction<(props: Record<string, unknown>, deviceMode: Device) => T>
-
-type DeserializedConfig<T> =
-  | T
-  | DeserializedFunction<(props: Record<string, unknown>, deviceMode: Device) => T>
-
-type GapXControlConfig = {
-  preset?: GapXControlValue
-  label?: string
-  defaultValue?: Gap
-  min?: number
-  max?: number
-  step?: number
-  hidden?: boolean
-}
-
-type SerializedGapXControl<_T = GapXControlValue> = {
-  type: typeof Controls.Types.GapX
-  options: SerializedConfig<GapXControlConfig>
-}
-
 function serializeGapXControl(control: GapXControl): [SerializedGapXControl, Transferable[]] {
   const { options } = control
 
@@ -314,11 +286,6 @@ function serializeGapXControl(control: GapXControl): [SerializedGapXControl, Tra
   const serializedOptions = serializeFunction(options)
 
   return [{ ...control, options: serializedOptions }, [serializedOptions]]
-}
-
-type DeserializedGapXControl<_T = GapXControlValue> = {
-  type: typeof Controls.Types.GapX
-  options: DeserializedConfig<GapXControlConfig>
 }
 
 function deserializeGapXControl(serializedControl: SerializedGapXControl): DeserializedGapXControl {
@@ -331,21 +298,6 @@ function deserializeGapXControl(serializedControl: SerializedGapXControl): Deser
   return { ...serializedControl, options: deserializedOptions }
 }
 
-type GapYControlConfig = {
-  preset?: GapYControlValue
-  label?: string
-  defaultValue?: Gap
-  min?: number
-  max?: number
-  step?: number
-  hidden?: boolean
-}
-
-type SerializedGapYControl<_T = GapYControlValue> = {
-  type: typeof Controls.Types.GapY
-  options: SerializedConfig<GapYControlConfig>
-}
-
 function serializeGapYControl(control: GapYControl): [SerializedGapYControl, Transferable[]] {
   const { options } = control
 
@@ -356,11 +308,6 @@ function serializeGapYControl(control: GapYControl): [SerializedGapYControl, Tra
   return [{ ...control, options: serializedOptions }, [serializedOptions]]
 }
 
-type DeserializedGapYControl<_T = GapYControlValue> = {
-  type: typeof Controls.Types.GapY
-  options: DeserializedConfig<GapYControlConfig>
-}
-
 function deserializeGapYControl(serializedControl: SerializedGapYControl): DeserializedGapYControl {
   const { options } = serializedControl
 
@@ -369,21 +316,6 @@ function deserializeGapYControl(serializedControl: SerializedGapYControl): Deser
   const deserializedOptions = deserializeFunction(options)
 
   return { ...serializedControl, options: deserializedOptions }
-}
-
-type ResponsiveNumberControlConfig = {
-  preset?: ResponsiveNumberControlValue
-  label?: string
-  defaultValue?: number
-  min?: number
-  max?: number
-  step?: number
-  hidden?: boolean
-}
-
-type SerializedResponsiveNumberControl<_T = ResponsiveNumberControlValue> = {
-  type: typeof Controls.Types.ResponsiveNumber
-  options: SerializedConfig<ResponsiveNumberControlConfig>
 }
 
 function serializeResponsiveNumberControl(
@@ -398,11 +330,6 @@ function serializeResponsiveNumberControl(
   return [{ ...control, options: serializedOptions }, [serializedOptions]]
 }
 
-type DeserializedResponsiveNumberControl<_T = ResponsiveNumberControlValue> = {
-  type: typeof Controls.Types.ResponsiveNumber
-  options: DeserializedConfig<ResponsiveNumberControlConfig>
-}
-
 function deserializeResponsiveNumberControl(
   serializedControl: SerializedResponsiveNumberControl,
 ): DeserializedResponsiveNumberControl {
@@ -413,17 +340,6 @@ function deserializeResponsiveNumberControl(
   const deserializedOptions = deserializeFunction(options)
 
   return { ...serializedControl, options: deserializedOptions }
-}
-
-type CheckboxControlConfig = {
-  preset?: CheckboxControlValue
-  label: string
-  hidden?: boolean
-}
-
-type SerializedCheckboxControl<_T = CheckboxControlValue> = {
-  type: typeof Controls.Types.Checkbox
-  options: SerializedConfig<CheckboxControlConfig>
 }
 
 function serializeCheckboxControl(
@@ -438,11 +354,6 @@ function serializeCheckboxControl(
   return [{ ...control, options: serializedOptions }, [serializedOptions]]
 }
 
-type DeserializedCheckboxControl<_T = CheckboxControlValue> = {
-  type: typeof Controls.Types.Checkbox
-  options: DeserializedConfig<CheckboxControlConfig>
-}
-
 function deserializeCheckboxControl(
   serializedControl: SerializedCheckboxControl,
 ): DeserializedCheckboxControl {
@@ -453,13 +364,6 @@ function deserializeCheckboxControl(
   const deserializedOptions = deserializeFunction(options)
 
   return { ...serializedControl, options: deserializedOptions }
-}
-
-type ResponsiveColorControlConfig = { label?: string; placeholder?: string; hidden?: boolean }
-
-type SerializedResponsiveColorControl<_T = ResponsiveColorControlValue> = {
-  type: typeof Controls.Types.ResponsiveColor
-  options: SerializedConfig<ResponsiveColorControlConfig>
 }
 
 function serializeResponsiveColorControl(
@@ -474,11 +378,6 @@ function serializeResponsiveColorControl(
   return [{ ...control, options: serializedOptions }, [serializedOptions]]
 }
 
-type DeserializedResponsiveColorControl<_T = ResponsiveColorControlValue> = {
-  type: typeof Controls.Types.ResponsiveColor
-  options: DeserializedConfig<ResponsiveColorControlConfig>
-}
-
 function deserializeResponsiveColorControl(
   serializedControl: SerializedResponsiveColorControl,
 ): DeserializedResponsiveColorControl {
@@ -489,21 +388,6 @@ function deserializeResponsiveColorControl(
   const deserializedOptions = deserializeFunction(options)
 
   return { ...serializedControl, options: deserializedOptions }
-}
-type NumberControlConfig = {
-  preset?: NumberControlValue
-  label?: string
-  defaultValue?: number
-  min?: number
-  max?: number
-  step?: number
-  suffix?: string
-  hidden?: boolean
-}
-
-type SerializedNumberControl<_T = NumberControlValue> = {
-  type: typeof Controls.Types.Number
-  options: SerializedConfig<NumberControlConfig>
 }
 
 function serializeNumberControl(control: NumberControl): [SerializedNumberControl, Transferable[]] {
@@ -516,11 +400,6 @@ function serializeNumberControl(control: NumberControl): [SerializedNumberContro
   return [{ ...control, options: serializedOptions }, [serializedOptions]]
 }
 
-type DeserializedNumberControl<_T = NumberControlValue> = {
-  type: typeof Controls.Types.Number
-  options: DeserializedConfig<NumberControlConfig>
-}
-
 function deserializeNumberControl(
   serializedControl: SerializedNumberControl,
 ): DeserializedNumberControl {
@@ -531,17 +410,6 @@ function deserializeNumberControl(
   const deserializedOptions = deserializeFunction(options)
 
   return { ...serializedControl, options: deserializedOptions }
-}
-type ResponsiveIconRadioGroupControlConfig<T extends string = string, U extends T = T> = {
-  label?: string
-  options: IconRadioGroupOption<T>[]
-  defaultValue?: U
-  hidden?: boolean
-}
-
-type SerializedResponsiveIconRadioGroupControl<_T = ResponsiveIconRadioGroupControlValue> = {
-  type: typeof Controls.Types.ResponsiveIconRadioGroup
-  options: SerializedConfig<ResponsiveIconRadioGroupControlConfig>
 }
 
 function serializeResponsiveIconRadioGroupControl(
@@ -556,11 +424,6 @@ function serializeResponsiveIconRadioGroupControl(
   return [{ ...control, options: serializedOptions }, [serializedOptions]]
 }
 
-type DeserializedResponsiveIconRadioGroupControl<_T = ResponsiveIconRadioGroupControlValue> = {
-  type: typeof Controls.Types.ResponsiveIconRadioGroup
-  options: DeserializedConfig<ResponsiveIconRadioGroupControlConfig>
-}
-
 function deserializeResponsiveIconRadioGroupControl(
   serializedControl: SerializedResponsiveIconRadioGroupControl,
 ): DeserializedResponsiveIconRadioGroupControl {
@@ -573,13 +436,6 @@ function deserializeResponsiveIconRadioGroupControl(
   return { ...serializedControl, options: deserializedOptions }
 }
 
-type DateControlConfig = { preset?: DateControlValue }
-
-type SerializedDateControl<_T = DateControlValue> = {
-  type: typeof Controls.Types.Date
-  options: SerializedConfig<DateControlConfig>
-}
-
 function serializeDateControl(control: DateControl): [SerializedDateControl, Transferable[]] {
   const { options } = control
 
@@ -588,11 +444,6 @@ function serializeDateControl(control: DateControl): [SerializedDateControl, Tra
   const serializedOptions = serializeFunction(options)
 
   return [{ ...control, options: serializedOptions }, [serializedOptions]]
-}
-
-type DeserializedDateControl<_T = DateControlValue> = {
-  type: typeof Controls.Types.Date
-  options: DeserializedConfig<DateControlConfig>
 }
 
 function deserializeDateControl(serializedControl: SerializedDateControl): DeserializedDateControl {
@@ -605,19 +456,6 @@ function deserializeDateControl(serializedControl: SerializedDateControl): Deser
   return { ...serializedControl, options: deserializedOptions }
 }
 
-type LinkControlConfig = {
-  preset?: LinkControlValue
-  label?: string
-  defaultValue?: LinkControlValue
-  options?: { value: LinkControlValue['type']; label: string }[]
-  hidden?: boolean
-}
-
-type SerializedLinkControl<_T = LinkControlValue> = {
-  type: typeof Controls.Types.Link
-  options: SerializedConfig<LinkControlConfig>
-}
-
 function serializeLinkControl(control: LinkControl): [SerializedLinkControl, Transferable[]] {
   const { options } = control
 
@@ -628,11 +466,6 @@ function serializeLinkControl(control: LinkControl): [SerializedLinkControl, Tra
   return [{ ...control, options: serializedOptions }, [serializedOptions]]
 }
 
-type DeserializedLinkControl<_T = LinkControlValue> = {
-  type: typeof Controls.Types.Link
-  options: DeserializedConfig<LinkControlConfig>
-}
-
 function deserializeLinkControl(serializedControl: SerializedLinkControl): DeserializedLinkControl {
   const { options } = serializedControl
 
@@ -641,13 +474,6 @@ function deserializeLinkControl(serializedControl: SerializedLinkControl): Deser
   const deserializedOptions = deserializeFunction(options)
 
   return { ...serializedControl, options: deserializedOptions }
-}
-
-type TextInputControlConfig = { label?: string; placeholder?: string; hidden?: boolean }
-
-type SerializedTextInputControl<_T = TextInputControlValue> = {
-  type: typeof Controls.Types.TextInput
-  options: SerializedConfig<TextInputControlConfig>
 }
 
 function serializeTextInputControl(
@@ -662,11 +488,6 @@ function serializeTextInputControl(
   return [{ ...control, options: serializedOptions }, [serializedOptions]]
 }
 
-type DeserializedTextInputControl<_T = TextInputControlValue> = {
-  type: typeof Controls.Types.TextInput
-  options: DeserializedConfig<TextInputControlConfig>
-}
-
 function deserializeTextInputControl(
   serializedControl: SerializedTextInputControl,
 ): DeserializedTextInputControl {
@@ -677,11 +498,6 @@ function deserializeTextInputControl(
   const deserializedOptions = deserializeFunction(options)
 
   return { ...serializedControl, options: deserializedOptions }
-}
-
-type SerializedResponsiveSelectControl<_T = ResponsiveSelectControlValue> = {
-  type: typeof Controls.Types.ResponsiveSelect
-  options: SerializedConfig<ResponsiveSelectControlConfig>
 }
 
 function serializeResponsiveSelectControl(
@@ -696,19 +512,6 @@ function serializeResponsiveSelectControl(
   return [{ ...control, options: serializedOptions }, [serializedOptions]]
 }
 
-type ResponsiveSelectControlConfig<T extends string = string, U extends T = T> = {
-  label?: string
-  labelOrientation?: SelectLabelOrientation
-  options: SelectOption<T>[]
-  defaultValue?: U
-  hidden?: boolean
-}
-
-type DeserializedResponsiveSelectControl<_T = ResponsiveSelectControlValue> = {
-  type: typeof Controls.Types.ResponsiveSelect
-  options: DeserializedConfig<ResponsiveSelectControlConfig>
-}
-
 function deserializeResponsiveSelectControl(
   serializedControl: SerializedResponsiveSelectControl,
 ): DeserializedResponsiveSelectControl {
@@ -719,18 +522,6 @@ function deserializeResponsiveSelectControl(
   const deserializedOptions = deserializeFunction(options)
 
   return { ...serializedControl, options: deserializedOptions }
-}
-
-type ResponsiveLengthControlConfig = {
-  label?: string
-  options?: LengthOption[]
-  defaultValue?: Length
-  hidden?: boolean
-}
-
-type SerializedResponsiveLengthControl<_T = ResponsiveLengthControlValue> = {
-  type: typeof Controls.Types.ResponsiveLength
-  options: SerializedConfig<ResponsiveLengthControlConfig>
 }
 
 function serializeResponsiveLengthControl(
@@ -745,11 +536,6 @@ function serializeResponsiveLengthControl(
   return [{ ...control, options: serializedOptions }, [serializedOptions]]
 }
 
-type DeserializedResponsiveLengthControl<_T = ResponsiveLengthControlValue> = {
-  type: typeof Controls.Types.ResponsiveLength
-  options: DeserializedConfig<ResponsiveLengthControlConfig>
-}
-
 function deserializeResponsiveLengthControl(
   serializedControl: SerializedResponsiveLengthControl,
 ): DeserializedResponsiveLengthControl {
@@ -760,11 +546,6 @@ function deserializeResponsiveLengthControl(
   const deserializedOptions = deserializeFunction(options)
 
   return { ...serializedControl, options: deserializedOptions }
-}
-
-type SerializedTextStyleControl<_T = TextStyleControlValue> = {
-  type: typeof Controls.Types.TextStyle
-  options: SerializedConfig<TextStyleControlConfig>
 }
 
 function serializeTextStyleControl(
@@ -779,17 +560,6 @@ function serializeTextStyleControl(
   return [{ ...control, options: serializedOptions }, [serializedOptions]]
 }
 
-type TextStyleControlConfig = {
-  preset?: TextStyleControlValue
-  label?: string
-  hidden?: boolean
-}
-
-type DeserializedTextStyleControl<_T = TextStyleControlValue> = {
-  type: typeof Controls.Types.TextStyle
-  options: DeserializedConfig<TextStyleControlConfig>
-}
-
 function deserializeTextStyleControl(
   serializedControl: SerializedTextStyleControl,
 ): DeserializedTextStyleControl {
@@ -802,11 +572,43 @@ function deserializeTextStyleControl(
   return { ...serializedControl, options: deserializedOptions }
 }
 
-type ImageControlConfig = { label?: string; hidden?: boolean }
+function serializeStyleV2Control(
+  definition: StyleV2ControlDefinition,
+): [Serialize<StyleV2ControlDefinition>, Transferable[]] {
+  const [type, transferables] = serializeControl(definition.config.type)
+  const getStyle = definition.config.getStyle && serializeFunction(definition.config.getStyle)
 
-type SerializedImageControl<_T = ImageControlValue> = {
-  type: typeof Controls.Types.Image
-  options: SerializedConfig<ImageControlConfig>
+  if (getStyle) {
+    transferables.push(getStyle)
+  }
+
+  return [
+    {
+      ...definition,
+      config: {
+        ...definition.config,
+        type,
+        getStyle,
+      },
+    },
+    transferables,
+  ] as [Serialize<StyleV2ControlDefinition>, Transferable[]]
+}
+
+function deserializeStyleV2Control(
+  definition: Serialize<StyleV2ControlDefinition>,
+): Deserialize<Serialize<StyleV2ControlDefinition>> {
+  const type = deserializeControl(definition.config.type)
+  const getStyle = definition.config.getStyle && deserializeFunction(definition.config.getStyle)
+
+  return {
+    ...definition,
+    config: {
+      ...definition.config,
+      getStyle,
+      type,
+    },
+  } as Deserialize<Serialize<StyleV2ControlDefinition>>
 }
 
 function serializeImageControl(control: ImageControl): [SerializedImageControl, Transferable[]] {
@@ -817,11 +619,6 @@ function serializeImageControl(control: ImageControl): [SerializedImageControl, 
   const serializedOptions = serializeFunction(options)
 
   return [{ ...control, options: serializedOptions }, [serializedOptions]]
-}
-
-type DeserializedImageControl<_T = ImageControlValue> = {
-  type: typeof Controls.Types.Image
-  options: DeserializedConfig<ImageControlConfig>
 }
 
 function deserializeImageControl(
@@ -836,11 +633,92 @@ function deserializeImageControl(
   return { ...serializedControl, options: deserializedOptions }
 }
 
-type RichTextControlConfig = { preset?: RichTextControlValue }
+export function serializeRichTextControlV2(
+  definition: RichTextV2ControlDefinition,
+): [Serialize<RichTextV2ControlDefinition>, Transferable[]] {
+  const { plugins, ...config } = definition.config
+  const transferables: Transferable[] = []
 
-type SerializedRichTextControl<_T = RichTextControlValue> = {
-  type: typeof Controls.Types.RichText
-  options: SerializedConfig<RichTextControlConfig>
+  return [
+    {
+      ...definition,
+      config: {
+        ...config,
+        plugins:
+          plugins?.map(plugin => {
+            const onKeyDown = plugin.onKeyDown && serializeFunction(plugin.onKeyDown)
+            if (onKeyDown) transferables.push(onKeyDown)
+            const withPlugin = plugin.withPlugin && serializeFunction(plugin.withPlugin)
+            if (withPlugin) transferables.push(withPlugin)
+
+            if (plugin.control) {
+              const [definition, pluginTransferables] = serializeControl(plugin.control.definition)
+              transferables.push(...pluginTransferables)
+              const getValue = serializeFunction(plugin.control.getValue)
+              if (getValue) transferables.push(getValue)
+              const onChange = serializeFunction(plugin.control.onChange)
+              if (onChange) transferables.push(onChange)
+
+              return {
+                control: {
+                  definition,
+                  onChange,
+                  getValue,
+                },
+                onKeyDown,
+                withPlugin,
+              }
+            }
+
+            return {
+              onKeyDown,
+              withPlugin,
+            }
+          }) ?? [],
+      },
+    },
+    transferables,
+  ] as [Serialize<RichTextV2ControlDefinition>, Transferable[]]
+}
+
+export function deserializeRichTextControlV2(
+  definition: Serialize<RichTextV2ControlDefinition>,
+): Deserialize<Serialize<RichTextV2ControlDefinition>> {
+  return {
+    ...definition,
+    config: {
+      ...definition.config,
+      plugins:
+        definition.config.plugins?.map(plugin => {
+          // TODO: you shouldn't need to cast all the functions
+          // There is a type missmatch where functions within the plugin aren't typed as Serialize<T>
+          // This prevents them from being used with `deserializeFunction`
+          const onKeyDown = plugin.onKeyDown && deserializeFunction(plugin.onKeyDown as any)
+          const withPlugin = plugin.withPlugin && deserializeFunction(plugin.withPlugin as any)
+
+          if (plugin.control) {
+            const definition = deserializeControl(plugin.control.definition as any)
+            const getValue = deserializeFunction(plugin.control.getValue as any)
+            const onChange = deserializeFunction(plugin.control.onChange as any)
+
+            return {
+              control: {
+                definition,
+                onChange,
+                getValue,
+              },
+              onKeyDown,
+              withPlugin,
+            }
+          }
+
+          return {
+            onKeyDown,
+            withPlugin,
+          }
+        }) ?? [],
+    },
+  } as Deserialize<Serialize<RichTextV2ControlDefinition>>
 }
 
 function serializeRichTextControl(
@@ -855,11 +733,6 @@ function serializeRichTextControl(
   return [{ ...control, options: serializedOptions }, [serializedOptions]]
 }
 
-type DeserializedRichTextControl<_T = RichTextControlValue> = {
-  type: typeof Controls.Types.RichText
-  options: DeserializedConfig<RichTextControlConfig>
-}
-
 function deserializeRichTextControl(
   serializedControl: SerializedRichTextControl,
 ): DeserializedRichTextControl {
@@ -871,132 +744,6 @@ function deserializeRichTextControl(
 
   return { ...serializedControl, options: deserializedOptions }
 }
-
-export type SerializedControl<T extends Data = Data> =
-  | Exclude<
-      Control<T>,
-      | ListControl<T extends ListControlValue ? T : ListControlValue>
-      | ShapeControl<T extends ShapeControlValue ? T : ShapeControlValue, any>
-      | TypeaheadControl<T extends TypeaheadControlValue ? T : TypeaheadControlValue>
-      | GapXControl<T>
-      | GapYControl<T>
-      | ResponsiveNumberControl<T>
-      | CheckboxControl<T>
-      | ResponsiveColorControl<T>
-      | NumberControl<T>
-      | ResponsiveIconRadioGroupControl<
-          T extends ResponsiveIconRadioGroupControlValue ? T : ResponsiveIconRadioGroupControlValue
-        >
-      | ResponsiveSelectControl<
-          T extends ResponsiveSelectControlValue ? T : ResponsiveSelectControlValue
-        >
-      | ResponsiveLengthControl<T>
-      | DateControl<T>
-      | LinkControl<T>
-      | TextInputControl<T>
-      | TextStyleControl<T>
-      | ImageControl<T>
-      | RichTextControl<T>
-      | RichTextV2ControlDefinition
-      | ComboboxControlDefinition
-      | ShapeControlDefinition
-      | ListControlDefinition
-      | StyleV2ControlDefinition
-    >
-  | SerializedListControl<T extends ListControlValue ? T : ListControlValue>
-  | SerializedShapeControl<T extends ShapeControlValue ? T : ShapeControlValue, any>
-  | SerializedTypeaheadControl<T extends TypeaheadControlValue ? T : TypeaheadControlValue>
-  | SerializedGapXControl<T>
-  | SerializedGapYControl<T>
-  | SerializedResponsiveNumberControl<T>
-  | SerializedCheckboxControl<T>
-  | SerializedResponsiveColorControl<T>
-  | SerializedNumberControl<T>
-  | SerializedResponsiveIconRadioGroupControl<T>
-  | SerializedResponsiveSelectControl<T>
-  | SerializedResponsiveLengthControl<T>
-  | SerializedDateControl<T>
-  | SerializedLinkControl<T>
-  | SerializedTextInputControl<T>
-  | SerializedTextStyleControl<T>
-  | SerializedImageControl<T>
-  | SerializedRichTextControl<T>
-  | Serialize<RichTextV2ControlDefinition>
-  | Serialize<ComboboxControlDefinition>
-  | Serialize<ShapeControlDefinition>
-  | Serialize<ListControlDefinition>
-  | Serialize<StyleV2ControlDefinition>
-
-type SerializedPanelControl<T extends Data = Data> = Extract<
-  SerializedControl<T>,
-  { type: PanelControlType }
->
-
-type SerializedPanelControlValueType<T extends SerializedPanelControl> =
-  T extends SerializedPanelControl<infer U> ? U : never
-
-export type DeserializedControl<T extends Data = Data> =
-  | Exclude<
-      Control<T>,
-      | ListControl<T extends ListControlValue ? T : ListControlValue>
-      | ShapeControl<T extends ShapeControlValue ? T : ShapeControlValue, any>
-      | TypeaheadControl<T extends TypeaheadControlValue ? T : TypeaheadControlValue>
-      | GapXControl<T>
-      | GapYControl<T>
-      | ResponsiveNumberControl<T>
-      | CheckboxControl<T>
-      | ResponsiveColorControl<T>
-      | NumberControl<T>
-      | ResponsiveIconRadioGroupControl<
-          T extends ResponsiveIconRadioGroupControlValue ? T : ResponsiveIconRadioGroupControlValue
-        >
-      | ResponsiveSelectControl<
-          T extends ResponsiveSelectControlValue ? T : ResponsiveSelectControlValue
-        >
-      | ResponsiveLengthControl<T>
-      | DateControl<T>
-      | LinkControl<T>
-      | TextInputControl<T>
-      | TextStyleControl<T>
-      | ImageControl<T>
-      | RichTextControl<T>
-      | RichTextV2ControlDefinition
-      | ComboboxControlDefinition
-      | ShapeControlDefinition
-      | ListControlDefinition
-      | StyleV2ControlDefinition
-    >
-  | DeserializedListControl<T extends ListControlValue ? T : ListControlValue>
-  | DeserializedShapeControl<T extends ShapeControlValue ? T : ShapeControlValue, any>
-  | DeserializedTypeaheadControl<T extends TypeaheadControlValue ? T : TypeaheadControlValue>
-  | DeserializedGapXControl<T>
-  | DeserializedGapYControl<T>
-  | DeserializedResponsiveNumberControl<T>
-  | DeserializedCheckboxControl<T>
-  | DeserializedResponsiveColorControl<T>
-  | DeserializedNumberControl<T>
-  | DeserializedResponsiveIconRadioGroupControl<T>
-  | DeserializedResponsiveSelectControl<T>
-  | DeserializedResponsiveLengthControl<T>
-  | DeserializedDateControl<T>
-  | DeserializedLinkControl<T>
-  | DeserializedTextInputControl<T>
-  | DeserializedTextStyleControl<T>
-  | DeserializedImageControl<T>
-  | DeserializedRichTextControl<T>
-  | Deserialize<Serialize<RichTextV2ControlDefinition>>
-  | Deserialize<Serialize<ComboboxControlDefinition>>
-  | Deserialize<Serialize<ShapeControlDefinition>>
-  | Deserialize<Serialize<ListControlDefinition>>
-  | Deserialize<Serialize<StyleV2ControlDefinition>>
-
-export type DeserializedPanelControl<T extends Data = Data> = Extract<
-  DeserializedControl<T>,
-  { type: PanelControlType }
->
-
-type DeserializedPanelControlValueType<T extends DeserializedPanelControl> =
-  T extends DeserializedPanelControl<infer U> ? U : never
 
 export function serializeControl<T extends Data>(
   control: Control<T>,
