+++ packages/runtime/src/builder/serialization/index.ts
@@ -1,3 +1,3 @@
-export * from './control-serialization'
-export * from './function-serialization'
+export { serializeControls, deserializeControls } from './control-serialization'
 export type { Serialize, Deserialize } from './controls/types'
+export type { SerializedControl } from './types'
